the only commit

+++ src/lib/Post.svelte
@@ -353,42 +353,50 @@
 							}}
 						/>
 						{#if post.user === $user.name || (post.post_origin === $chat._id && $chat.owner === $user.name)}
-							<button
-								class="circle trash"
-								bind:this={deleteButton}
-								on:click={async () => {
-									if (shiftHeld) {
-										deleteButton.disabled = true;
-										try {
-											const resp = await fetch(
-												`${apiUrl}posts?id=${post.post_id}`,
-												{
-													method: "DELETE",
-													headers: $authHeader,
-												}
-											);
-											if (!resp.ok) {
-												if (resp.status === 429) {
+							{#if !bridged}
+								<button
+									class="circle trash"
+									bind:this={deleteButton}
+									on:click={async () => {
+										if (shiftHeld) {
+											deleteButton.disabled = true;
+											try {
+												const resp = await fetch(
+													`${apiUrl}posts?id=${post.post_id}`,
+													{
+														method: "DELETE",
+														headers: $authHeader,
+													}
+												);
+												if (!resp.ok) {
+													if (resp.status === 429) {
+														throw new Error(
+															"Too many requests! Try again later."
+														);
+													}
 													throw new Error(
-														"Too many requests! Try again later."
+														"Response code is not OK; code is " +
+															resp.status
 													);
 												}
-												throw new Error(
-													"Response code is not OK; code is " +
-														resp.status
-												);
+											} catch (e) {
+												editError = e;
 											}
-										} catch (e) {
-											editError = e;
+											deleteButton.disabled = false;
+										} else {
+											modals.showModal(DeletePostModal, {
+												post,
+											});
 										}
-										deleteButton.disabled = false;
-									} else {
-										modals.showModal(DeletePostModal, {
-											post,
-										});
-									}
-								}}
-							/>
+									}}
+								/>
+							{:else}
+								<button
+									class="circle trash"
+									disabled
+									title="You cannot delete bridged messages."
+								/>
+							{/if}
 						{:else}
 							<button
 								class="circle report"
